refactored component's main-page into 'index'
