+++ server/client/src/components/channel/dashboard/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import {connect} from 'react-redux';
 
-import './Dashboard.css';
+import './index.css';
 
 class DashBoard extends Component {
   render(){
